pyproject.toml
[project]
name = "resume-builder-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "dj-rest-auth>=5.0.2",
    "django>=6.0",
    "django-allauth[socialaccount]>=65.13.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "environ>=1.0",
    "factory-boy>=3.3.3",
    "markdown>=3.10",
    "openai>=2.9.0",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pypdf2>=3.0.1",
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "ruff>=0.15.0",
    "weasyprint>=67.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
